backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
boto3>=1.34.0
pydantic>=2.5.0
python-multipart>=0.0.9
Pillow>=10.0.0
numpy>=1.24.0

# Optionnel si vous utilisez un modèle ONNX
onnxruntime>=1.16.0

# Optionnel si vous chargez un .ckpt via ml/src (PyTorch)
# torch, torchvision, lightning, albumentations, omegaconf, torchmetrics, wandb
# torch>=2.2.0
# torchvision>=0.17.0
# lightning>=2.2.0
# albumentations>=1.3.0
# omegaconf>=2.3.0
# torchmetrics>=1.3.0
# wandb>=0.16.0

